Added Policy-Driven Authorization

+++ app/Http/Controllers/Admin/CategoryController.php
@@ -3,8 +3,10 @@
 namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
-use BalajiDharma\LaravelAdminCore\Requests\StoreCategoryRequest;
-use BalajiDharma\LaravelAdminCore\Requests\UpdateCategoryRequest;
+use BalajiDharma\LaravelAdminCore\Actions\Category\CategoryCreateAction;
+use BalajiDharma\LaravelAdminCore\Actions\Category\CategoryUpdateAction;
+use BalajiDharma\LaravelAdminCore\Data\Category\CategoryCreateData;
+use BalajiDharma\LaravelAdminCore\Data\Category\CategoryUpdateData;
 use BalajiDharma\LaravelCategory\Models\Category;
 use BalajiDharma\LaravelCategory\Models\CategoryType;
 use Illuminate\Support\Facades\Auth;
@@ -12,14 +14,6 @@ use Inertia\Inertia;
 
 class CategoryController extends Controller
 {
-    public function __construct()
-    {
-        $this->middleware('can:category list', ['only' => ['index', 'show']]);
-        $this->middleware('can:category create', ['only' => ['create', 'store']]);
-        $this->middleware('can:category edit', ['only' => ['edit', 'update']]);
-        $this->middleware('can:category delete', ['only' => ['destroy']]);
-    }
-
     /**
      * Display a listing of the resource.
      *
@@ -27,6 +21,7 @@ class CategoryController extends Controller
      */
     public function index(CategoryType $type)
     {
+        $this->authorize('adminViewAny', Category::class);
         $items = (new Category)->toTree($type->id, true);
 
         return Inertia::render('Admin/Category/Item/Index', [
@@ -47,6 +42,7 @@ class CategoryController extends Controller
      */
     public function create(CategoryType $type)
     {
+        $this->authorize('adminCreate', Category::class);
         $itemOptions = Category::selectOptions($type->id, null, true);
 
         return Inertia::render('Admin/Category/Item/Create', [
@@ -60,13 +56,10 @@ class CategoryController extends Controller
      *
      * @return \Illuminate\Http\RedirectResponse
      */
-    public function store(StoreCategoryRequest $request, CategoryType $type)
+    public function store(CategoryCreateData $data, CategoryType $type, CategoryCreateAction $categoryCreateAction)
     {
-        if (! $request->has('enabled')) {
-            $request['enabled'] = false;
-        }
-
-        $type->categories()->create($request->all());
+        $this->authorize('adminCreate', Category::class);
+        $categoryCreateAction->handle($data, $type);
 
         return redirect()->route('admin.category.type.item.index', $type->id)
             ->with('message', 'Category created successfully.');
@@ -79,6 +72,7 @@ class CategoryController extends Controller
      */
     public function edit(CategoryType $type, Category $item)
     {
+        $this->authorize('adminUpdate', $item);
         $itemOptions = Category::selectOptions($type->id, $item->parent_id ?? $item->id);
 
         return Inertia::render('Admin/Category/Item/Edit', [
@@ -93,13 +87,10 @@ class CategoryController extends Controller
      *
      * @return \Illuminate\Http\RedirectResponse
      */
-    public function update(UpdateCategoryRequest $request, CategoryType $type, Category $item)
+    public function update(CategoryUpdateData $data, CategoryType $type, Category $item, CategoryUpdateAction $categoryUpdateAction)
     {
-        if (! $request->has('enabled')) {
-            $request['enabled'] = false;
-        }
-
-        $item->update($request->all());
+        $this->authorize('adminUpdate', $item);
+        $categoryUpdateAction->handle($data, $item);
 
         return redirect()->route('admin.category.type.item.index', $type->id)
             ->with('message', 'Category updated successfully.');
@@ -113,6 +104,7 @@ class CategoryController extends Controller
      */
     public function destroy(CategoryType $type, Category $item)
     {
+        $this->authorize('adminDelete', $item);
         $item->delete();
 
         return redirect()->route('admin.category.type.item.index', $type->id)
