+++ app/Http/Controllers/Admin/MenuController.php
@@ -3,22 +3,16 @@
 namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
-use BalajiDharma\LaravelAdminCore\Requests\StoreMenuRequest;
-use BalajiDharma\LaravelAdminCore\Requests\UpdateMenuRequest;
+use BalajiDharma\LaravelAdminCore\Actions\Menu\MenuCreateAction;
+use BalajiDharma\LaravelAdminCore\Actions\Menu\MenuUpdateAction;
+use BalajiDharma\LaravelAdminCore\Data\Menu\MenuCreateData;
+use BalajiDharma\LaravelAdminCore\Data\Menu\MenuUpdateData;
 use BalajiDharma\LaravelMenu\Models\Menu;
 use Illuminate\Support\Facades\Auth;
 use Inertia\Inertia;
 
 class MenuController extends Controller
 {
-    public function __construct()
-    {
-        $this->middleware('can:menu list', ['only' => ['index']]);
-        $this->middleware('can:menu create', ['only' => ['create', 'store']]);
-        $this->middleware('can:menu edit', ['only' => ['edit', 'update']]);
-        $this->middleware('can:menu delete', ['only' => ['destroy']]);
-    }
-
     /**
      * Display a listing of the resource.
      *
@@ -26,6 +20,7 @@ class MenuController extends Controller
      */
     public function index()
     {
+        $this->authorize('adminViewAny', Menu::class);
         $menus = (new Menu)->newQuery();
 
         if (request()->has('search')) {
@@ -67,6 +62,7 @@ class MenuController extends Controller
      */
     public function create()
     {
+        $this->authorize('adminCreate', Menu::class);
         return Inertia::render('Admin/Menu/Create');
     }
 
@@ -75,13 +71,10 @@ class MenuController extends Controller
      *
      * @return \Illuminate\Http\RedirectResponse
      */
-    public function store(StoreMenuRequest $request)
+    public function store(MenuCreateData $data, MenuCreateAction $menuCreateAction)
     {
-        Menu::create([
-            'name' => $request->name,
-            'machine_name' => $request->machine_name,
-            'description' => $request->description,
-        ]);
+        $this->authorize('adminCreate', Menu::class);
+        $menuCreateAction->handle($data);
 
         return redirect()->route('admin.menu.index')
             ->with('message', 'Menu created successfully.');
@@ -94,6 +87,7 @@ class MenuController extends Controller
      */
     public function edit(Menu $menu)
     {
+        $this->authorize('adminUpdate', $menu);
         return Inertia::render('Admin/Menu/Edit', [
             'menu' => $menu,
         ]);
@@ -104,9 +98,10 @@ class MenuController extends Controller
      *
      * @return \Illuminate\Http\RedirectResponse
      */
-    public function update(UpdateMenuRequest $request, Menu $menu)
+    public function update(MenuUpdateData $data, Menu $menu, MenuUpdateAction $menuUpdateAction)
     {
-        $menu->update($request->all());
+        $this->authorize('adminUpdate', $menu);
+        $menuUpdateAction->handle($data, $menu);
 
         return redirect()->route('admin.menu.index')
             ->with('message', 'Menu updated successfully.');
@@ -119,6 +114,7 @@ class MenuController extends Controller
      */
     public function destroy(Menu $menu)
     {
+        $this->authorize('adminDelete', $menu);
         $menu->delete();
 
         return redirect()->route('admin.menu.index')
